[MCSE-30] Use OSA_DisableIRQGlobal everywhere
The code is mixing the use of OSA_DisableIRQGlobal() and
OSA_InterruptDisable() which was luckily ok with OSA BM as it's the
same underlying API, but when using the OSA ThreadX port, the behavior
is different, so it's important to align which API is used.
As OSA BM was masking the PRIMASK, using OSA_DisableIRQGlobal() will
have the same effect.

Signed-off-by: Axel Le Bourhis <axel.lebourhis@nxp.com>