[LL-1726][HADM] SW Workaround for rare case where RSM does not abort mode0 timeout sequence
KW47 HW bug workaround on reflector: when mode0 retry is enabled, there's a small
probability that the RSM does not send an abort IRQ in case the last mode0 is
received (AA match) just after transitioning to 1st non_mode0 step.
In this case, the condition is detected by an RSM underrun on 2nd non-mode0 step.
This situation should be avoided by the LL ensuring that mode0 timeout
is only enabled when devices are properly synchronized (small window widening).