[LL-1862] [HADM] Re-sync DRBG in case of missed subevent
The HAL is responsible for consuming DRBG bits wether or not the
subevent was successful as only the HAL knows how many steps config have
been already generated.
This change has a dependency on LL change counterpart, which is the
possibility to pass a NULL pointer to BLE_HADM_DRBG_Generate_CS_SYNC_step().